# lib/api-zod/src/generated/types/contribution.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Podcast - Friends Hangout Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Contribution {
  id: number;
  userId: number;
  userName: string;
  /** @nullable */
  userAvatarUrl?: string | null;
  amount: number;
  updatedAt: string;
}