# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Podcast - Friends Hangout Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Message } from './message';
import type { Schedule } from './schedule';

export interface DashboardSummary {
  /** @nullable */
  leadingLocation: string | null;
  /** @nullable */
  leadingLocationId?: number | null;
  leadingLocationVotes?: number;
  totalVotes: number;
  schedule: Schedule;
  recentMessages: Message[];
  /** Sum of all contribution amounts */
  totalContributions: number;
  /** Count of active loans */
  activeLoans: number;
  unreadNotifications: number;
}