# lib/api-zod/src/generated/types/locationWithVotes.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Podcast - Friends Hangout Management Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface LocationWithVotes {
  id: number;
  name: string;
  voteCount: number;
}